test(storage): move the v2 test harness into main_test.go per dialect
TestMain, the shared pool var, and the cross-cutting project helpers lived in
project_test.go but are used by the team and json schema tests too. Move them
to a dedicated main_test.go in both the postgres and spanner v2 dialect
packages so the harness is where readers expect it and project_test.go holds
only project tests.
