+++ internal/storage/v2/dialect/spanner/project_test.go
@@ -5,12 +5,6 @@ package spanner
 import (
 	"context"
 	"crypto/rand"
-	"database/sql"
-	"fmt"
-	"log/slog"
-	"os"
-	"strconv"
-	"strings"
 	"testing"
 	"time"
 
@@ -18,85 +12,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/zitadel/nextgen/internal/domain"
 	"github.com/zitadel/nextgen/internal/storage/database"
-	"github.com/zitadel/nextgen/internal/storage/database/dbtest"
-	spannerdialect "github.com/zitadel/nextgen/internal/storage/database/dialect/spanner"
-	"github.com/zitadel/nextgen/internal/storage/database/dialect/spanner/migration"
 	v2database "github.com/zitadel/nextgen/internal/storage/v2/database"
 )
-
-// testClient is a v2 spanner client connected to the migrated test database,
-// shared across the tests in this package.
-var testClient *Client
-
-func TestMain(m *testing.M) {
-	code, err := run(m)
-	if err != nil {
-		slog.Error("spanner test setup failed", "error", err)
-	}
-	os.Exit(code)
-}
-
-func run(m *testing.M) (int, error) {
-	ctx := context.Background()
-
-	connector, stop, err := dbtest.Spanner(ctx)
-	if err != nil {
-		return 1, err
-	}
-	defer stop()
-
-	cfg, ok := connector.(*spannerdialect.Config)
-	if !ok {
-		return 1, fmt.Errorf("expected *spanner.Config connector, got %T", connector)
-	}
-
-	db, err := sql.Open("spanner", cfg.DSN)
-	if err != nil {
-		return 1, err
-	}
-	defer db.Close()
-	if err := migration.Migrate(ctx, db); err != nil {
-		return 1, err
-	}
-
-	dialect, err := DecodeConfig(cfg.DSN)
-	if err != nil {
-		return 1, err
-	}
-	pool, err := dialect.Connect(ctx)
-	if err != nil {
-		return 1, err
-	}
-	defer pool.Close(ctx)
-
-	c, ok := pool.(*Client)
-	if !ok {
-		return 1, fmt.Errorf("expected *Client from the v2 spanner dialect, got %T", pool)
-	}
-	testClient = c
-
-	return m.Run(), nil
-}
-
-// uniqueProjectID returns a collision-free project ID scoped to the running
-// (sub)test. The v2 statements commit immediately (no rollback), so isolation
-// relies on unique IDs plus DeleteProjectByID cleanup rather than a transaction.
-func uniqueProjectID(t *testing.T) string {
-	t.Helper()
-	return "proj-" + strings.ReplaceAll(t.Name(), "/", "_") + "-" + strconv.FormatInt(time.Now().UnixNano(), 10)
-}
-
-func newTestProject(id string) *domain.Project {
-	return &domain.Project{ID: id, Name: "project-" + rand.Text(), PreviewOrigins: []string{}}
-}
-
-func projectIDs(projects []*domain.Project) []string {
-	ids := make([]string, len(projects))
-	for i, p := range projects {
-		ids[i] = p.ID
-	}
-	return ids
-}
 
 func TestProjectStatements_Create(t *testing.T) {
 	stmts := testClient.Statements()
